Student class and header finished

+++ ex02-find-extremes/Student.h
@@ -2,8 +2,8 @@ namespace ds_course{
     class Student{
     public:
         Student();
-        int age;
-        double height;
+        unsigned long long int age;
+        long double height;
         int compareTo(Student ss);
     };
 }
